# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * All Window Door Parts API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductSpecifications = { [key: string]: string };

export interface Product {
  id: number;
  sku: string;
  name: string;
  description?: string;
  price: number;
  originalPrice?: number;
  category: string;
  subcategory?: string;
  supplier: string;
  inStock: boolean;
  imageUrl?: string;
  tags?: string[];
  specifications?: ProductSpecifications;
  compatibleBrands?: string[];
  createdAt?: string;
}

export interface ProductsResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description?: string;
  productCount: number;
  imageUrl?: string;
}

export interface PartsIdRequest {
  name: string;
  email: string;
  phone?: string;
  description: string;
  windowDoorBrand?: string;
  windowDoorAge?: string;
  imageBase64?: string;
  imageFileName?: string;
}

export interface PartsIdResponse {
  success: boolean;
  ticketId?: string;
  message: string;
}

export interface ContactRequest {
  name: string;
  email: string;
  phone?: string;
  subject?: string;
  message: string;
}

export interface ContactResponse {
  success: boolean;
  message: string;
}

export type CatalogStatsSupplierBreakdownItem = {
  supplier: string;
  count: number;
};

export type CatalogStatsTopCategoriesItem = {
  category: string;
  count: number;
};

export interface CatalogStats {
  totalProducts: number;
  totalCategories: number;
  supplierBreakdown: CatalogStatsSupplierBreakdownItem[];
  topCategories: CatalogStatsTopCategoriesItem[];
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GetProductsParams = {
  category?: string;
  search?: string;
  page?: number;
  limit?: number;
  sort?: string;
};

export type GetSearchSuggestionsParams = {
  q: string;
};