# lib/api-zod/src/generated/types/catalogStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * All Window Door Parts API
 * OpenAPI spec version: 0.1.0
 */
import type { CatalogStatsSupplierBreakdownItem } from "./catalogStatsSupplierBreakdownItem";
import type { CatalogStatsTopCategoriesItem } from "./catalogStatsTopCategoriesItem";

export interface CatalogStats {
  totalProducts: number;
  totalCategories: number;
  supplierBreakdown: CatalogStatsSupplierBreakdownItem[];
  topCategories: CatalogStatsTopCategoriesItem[];
}