# lib/api-zod/src/generated/types/contactRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * All Window Door Parts API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactRequest {
  name: string;
  email: string;
  phone?: string;
  subject?: string;
  message: string;
}