# lib/api-zod/src/generated/types/partsIdRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * All Window Door Parts API
 * OpenAPI spec version: 0.1.0
 */

export interface PartsIdRequest {
  name: string;
  email: string;
  phone?: string;
  description: string;
  windowDoorBrand?: string;
  windowDoorAge?: string;
  imageBase64?: string;
  imageFileName?: string;
}