# lib/api-zod/src/generated/types/partsIdResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * All Window Door Parts API
 * OpenAPI spec version: 0.1.0
 */

export interface PartsIdResponse {
  success: boolean;
  ticketId?: string;
  message: string;
}