# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * All Window Door Parts API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductSpecifications } from "./productSpecifications";

export interface Product {
  id: number;
  sku: string;
  name: string;
  description?: string;
  price: number;
  originalPrice?: number;
  category: string;
  subcategory?: string;
  supplier: string;
  inStock: boolean;
  imageUrl?: string;
  tags?: string[];
  specifications?: ProductSpecifications;
  compatibleBrands?: string[];
  createdAt?: Date;
}